adding validation to avoid empty inputs

+++ js/index.js
@@ -16,21 +16,30 @@ if (localStorage.getItem("product") !== null) {
 }
 
 function adding_Card() {
-  var array_Object = {
-    name: input_ProductName.value,
-    price: input_productPrice.value,
-    category: input_productCategory.value,
-    description: input_productDescription.value,
-    image: input_productImage.files[0]
-      ? `images/${input_productImage.files[0]?.name}`
-      : "images/2.jpg",
-  };
-  array_list.push(array_Object);
-  localStorage.setItem("product", JSON.stringify(array_list));
+  if (input_ProductName.value) {
+    var array_Object = {
+      name: input_ProductName.value,
+      price: input_productPrice.value,
+      category: input_productCategory.value,
+      description: input_productDescription.value,
+      image: input_productImage.files[0]
+        ? `images/${input_productImage.files[0]?.name}`
+        : "images/2.jpg",
+    };
+    array_list.push(array_Object);
+    localStorage.setItem("product", JSON.stringify(array_list));
 
-  display_Card();
-  // to clear the inputs after adding the card
-  clear_inputs();
+    display_Card();
+    // to clear the inputs after adding the card
+    clear_inputs();
+  } else {
+    Swal.fire({
+      icon: "error",
+      title: "Oops...",
+      text: "Something went wrong!",
+      footer: '<a href="#">Why do I have this issue?</a>',
+    });
+  }
 }
 
 function display_Card() {
@@ -68,9 +77,26 @@ function clear_inputs() {
 }
 
 function clear_item(index) {
-  array_list.splice(index, 1);
-  localStorage.setItem("product", JSON.stringify(array_list));
-  display_Card();
+  Swal.fire({
+    title: "Are you sure?",
+    text: "You won't be able to revert this!",
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonColor: "#3085d6",
+    cancelButtonColor: "#d33",
+    confirmButtonText: "Yes, delete it!",
+  }).then((result) => {
+    if (result.isConfirmed) {
+      array_list.splice(index, 1);
+      localStorage.setItem("product", JSON.stringify(array_list));
+      display_Card();
+      Swal.fire({
+        title: "Deleted!",
+        text: "Your file has been deleted.",
+        icon: "success",
+      });
+    }
+  });
 }
 function recale_Item(index) {
   input_ProductName.value = array_list[index].name;
